refactor: 폴더 구조 재정립 (#13)
- app 폴더는 라우팅만 담당, 기능은 feature에서 담당
- 공유되는 폴더들은 shared 내부로 이동
