--- emotion.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="@emotion/react/types/css-prop" />
