+++ features/note/components/AutoResizeTextarea.tsx
@@ -2,7 +2,7 @@
 
 import { useLayoutEffect, useRef } from "react";
 
-import { cn } from "@/lib/cn";
+import { cn } from "@/shared/lib/cn";
 
 import { MIN_TEXTAREA_ROWS } from "../constants";
