+++ shared/components/Sidebar/UserProfile.tsx
