+++ shared/components/__tests__/DarkModeToggle.test.tsx
@@ -2,7 +2,11 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
-import { THEME_COOKIE_KEY, THEME_DARK, THEME_LIGHT } from "@/constants/theme";
+import {
+  THEME_COOKIE_KEY,
+  THEME_DARK,
+  THEME_LIGHT,
+} from "@/shared/constants/theme";
 
 import { DarkModeToggle } from "../Sidebar/DarkModeToggle";
