+++ sonar-project.properties
@@ -2,7 +2,7 @@ sonar.projectKey=KimDongGyun23_deloki
 sonar.organization=kimdonggyun23
 
 sonar.sources=.
-sonar.exclusions=**/*.test.ts,**/*.test.tsx,**/node_modules/**,**/.next/**,**/coverage/**,app/**,constants/**
+sonar.exclusions=public/**,**/*.test.ts,**/*.test.tsx,**/node_modules/**,**/.next/**,**/coverage/**,app/**,**/constants/**
 
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
